Fix javascript loading for SSL embeds.
In production, liveupdate embeds are hosted on a off-cookie-domain. This
domain would also fail the is_reddit_url test which would mean that the base
URL wouldn't be set as secure even if the client connection was secure. This
would in turn cause javascript to not execute because it was loaded
insecurely.